spark/requirements.txt
tensorflow
tweet-preprocessor
nltk
pyspark
kafka-python